requirements.txt
PyYAML>=6.0,<7
Jinja2>=3.1,<4
pytest>=8.3,<9
